Introduce Queue/StackAdapter, use in FixedQueue/FixedStack
These are the equivalent to std::queue/std::stack but with
fixed_container properties.
