+++ CMakeLists.txt
@@ -189,8 +189,12 @@ if(BUILD_TESTS)
     add_test_dependencies(pair_test)
     add_executable(pair_view_test test/pair_view_test.cpp)
     add_test_dependencies(pair_view_test)
+    add_executable(queue_adapter_test test/queue_adapter_test.cpp)
+    add_test_dependencies(queue_adapter_test)
     add_executable(reflection_test test/reflection_test.cpp)
     add_test_dependencies(reflection_test)
+    add_executable(stack_adapter_test test/stack_adapter_test.cpp)
+    add_test_dependencies(stack_adapter_test)
     add_executable(string_literal_test test/string_literal_test.cpp)
     add_test_dependencies(string_literal_test)
     add_executable(type_name_test test/type_name_test.cpp)
