+++ include/fixed_containers/fixed_stack.hpp
@@ -3,6 +3,7 @@
 #include "fixed_containers/fixed_vector.hpp"
 #include "fixed_containers/sequence_container_checking.hpp"
 #include "fixed_containers/source_location.hpp"
+#include "fixed_containers/stack_adapter.hpp"
 
 #include <cstddef>
 
@@ -12,21 +13,14 @@ template <typename T,
           std::size_t MAXIMUM_SIZE,
           customize::SequenceContainerChecking CheckingType =
               customize::SequenceContainerAbortChecking<T, MAXIMUM_SIZE>>
-class FixedStack
+class FixedStack : public StackAdapter<FixedVector<T, MAXIMUM_SIZE, CheckingType>>
 {
-public:
-    using container_type = FixedVector<T, MAXIMUM_SIZE, CheckingType>;
-    using value_type = typename container_type::value_type;
-    using size_type = typename container_type::size_type;
-    using reference = typename container_type::reference;
-    using const_reference = typename container_type::const_reference;
-
-public:  // Public so this type is a structural type and can thus be used in template parameters
-    container_type IMPLEMENTATION_DETAIL_DO_NOT_USE_data_;
+private:
+    using Base = StackAdapter<FixedVector<T, MAXIMUM_SIZE, CheckingType>>;
 
 public:
     constexpr FixedStack()
-      : IMPLEMENTATION_DETAIL_DO_NOT_USE_data_{}
+      : Base{}
     {
     }
 
@@ -35,66 +29,8 @@ public:
                          InputIt last,
                          const std_transition::source_location& loc =
                              std_transition::source_location::current()) noexcept
-      : IMPLEMENTATION_DETAIL_DO_NOT_USE_data_(first, last, loc)
+      : Base{first, last, loc}
     {
-    }
-
-public:
-    [[nodiscard]] constexpr std::size_t max_size() const noexcept { return MAXIMUM_SIZE; }
-    [[nodiscard]] constexpr std::size_t size() const noexcept
-    {
-        return IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.size();
-    }
-    [[nodiscard]] constexpr bool empty() const noexcept { return size() == 0; }
-
-    constexpr reference top(
-        const std_transition::source_location& loc = std_transition::source_location::current())
-    {
-        return IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.back(loc);
-    }
-    constexpr const_reference top(const std_transition::source_location& loc =
-                                      std_transition::source_location::current()) const
-    {
-        return IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.back(loc);
-    }
-
-    constexpr void push(
-        const value_type& value,
-        const std_transition::source_location& loc = std_transition::source_location::current())
-    {
-        IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.push_back(value, loc);
-    }
-    constexpr void push(
-        value_type&& value,
-        const std_transition::source_location& loc = std_transition::source_location::current())
-    {
-        IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.push_back(std::move(value), loc);
-    }
-
-    template <class... Args>
-    constexpr decltype(auto) emplace(Args&&... args)
-    {
-        IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.emplace_back(std::forward<Args>(args)...);
-    }
-
-    constexpr void pop(
-        const std_transition::source_location& loc = std_transition::source_location::current())
-    {
-        IMPLEMENTATION_DETAIL_DO_NOT_USE_data_.pop_back(loc);
-    }
-
-    template <std::size_t MAXIMUM_SIZE_2, customize::SequenceContainerChecking CheckingType2>
-    constexpr bool operator==(const FixedStack<T, MAXIMUM_SIZE_2, CheckingType2>& other) const
-    {
-        return IMPLEMENTATION_DETAIL_DO_NOT_USE_data_ ==
-               other.IMPLEMENTATION_DETAIL_DO_NOT_USE_data_;
-    }
-
-    template <std::size_t MAXIMUM_SIZE_2, customize::SequenceContainerChecking CheckingType2>
-    constexpr auto operator<=>(const FixedStack<T, MAXIMUM_SIZE_2, CheckingType2>& other) const
-    {
-        return IMPLEMENTATION_DETAIL_DO_NOT_USE_data_ <=>
-               other.IMPLEMENTATION_DETAIL_DO_NOT_USE_data_;
     }
 };
